+++ main.ts
@@ -54,7 +54,7 @@ const BOT_ADMINS = ["334538784043696130", "412365502112071681"]
 
 let PREFIX = "["
 
-const VERSION = "1.5.2"
+const VERSION = "1.5.3"
 
 let SPAMS = []
 
@@ -87,7 +87,7 @@ echo:
         }
         return {
             reply: msgToReplyTo,
-            content: expandContent(this.content, msg)
+            content: expandContent(this.content, msg.author)
         }
     }, 
     'echo [-Df] [reply=\"messageid\"] [filename=\"name\"] [ext=\"ext\"] message\n-D: don\'t delete your message\n-f: write to file', 'Df').setCategory("fun").setMeta({version: "1.0.0"})
@@ -169,7 +169,7 @@ button:
                         collector.on("collect", async i => {
                             if(i.customId === `${msg.author.id}:button`){
                                 await i.update({
-                                    content: expandContent(onclick, msg, {
+                                    content: expandContent(onclick, msg.author, {
                                         "clicker": i.user.username,
                                         "clickerm": userMention(i.user.id),
                                         "timeclicked": new Date().toString()
@@ -698,7 +698,7 @@ snipe:
 "8ball":
     new Command(function(msg, opts){
         let resp = JSON.parse(fs.readFileSync('./storage/8ball.list').toString())
-        return {content: expandContent(resp[Math.floor(Math.random() * resp.length)], msg, {content: this.content.trim() || "?"})}
+        return {content: expandContent(resp[Math.floor(Math.random() * resp.length)], msg.author, {content: this.content.trim() || "?"})}
     }).setCategory("fun").setMeta({version: "1.2.0"})
 ,
 "8bfile": 
@@ -724,25 +724,46 @@ money:
     }, "money user").setCategory("economy").setMeta({version: "1.3.0"})
 ,
 profile:
-    new Command(function(msg, opts){
-        let user = msg.author.id
+    new Command(function(msg: Message, opts){
+        let u = msg.author
+        let fmt = this.getAttr("fmt") || "%i:\nmoney: %m\ntax rate: %t"
         if(this.content) {
-            let u = userFinder(msg.guild, this.content)
-            user = u.first()?.id
+            if(!(u = userFinder(msg.guild, this.content)?.first()?.user)) return {content: `${this.content} not found`}
         }
-        if(!user) return {content: `${this.content} not found`}
         if(opts["f"]){
             return {files: [{
-                attachment: `./storage/${user}.json`,
-                name: `${user}.json`
+                attachment: `./storage/${u.id}.json`,
+                name: `${u.id}.json`
             }]}
         }
-        let data = ""
-        for(let i in users[user]){
-            data += `${i}: ${users[user][i]}\n`
+        if(!users[u.id])return {content: "this user has no profile"}
+        fmt = expandContent(fmt, u)
+        return {
+            content: formatp(fmt, [
+                ["(?<!%)%i", u.id],
+                ["(?<!%)%m", String(users[u.id].money)],
+                ["(?<!%)%lt", new Date(users[u.id].lastTalked)],
+                ["(?<!%)%lx", new Date(users[u.id].lastTaxed)],
+                ["(?<!%)%ld", new Date(users[u.id].lastDonated)],
+                ["(?<!%)%t", String(users[u.id].taxRate)],
+                ["(?<!%)%v", (() => {
+                    let text = ""
+                    for(let va in users[u.id].vars){
+                        text += `${va}: ${users[u.id].vars[va]}\n`
+                    }
+                    return text || "NO VARS"
+                })],
+            ])
         }
-        return {content: data.trim()}
-    }).setCategory("economy").setMeta({version: "1.3.0"})
+    }, `profile [fmt=\"fmt\"] [user]
+formats:
+    %i: user id
+    %m: user money
+    %lt: last time user talked
+    %lx: last time user taxed someone
+    %ld: last time user donated
+    %t: user's taxrate
+    %v: user's variables`).setCategory("economy").setMeta({version: "1.3.0"})
 ,
 leaderboard:
     new Command(function(msg: Message, opts){
@@ -1205,6 +1226,7 @@ edititem:
     })
 }
 
+commands["allitems"].registerAlias(["shop"], commands)
 commands["calc"].registerAlias(["c", "eval", "evaluate"], commands)
 commands["leaderboard"].registerAlias(["lb", "top"], commands)
 commands["8bfile"].registerAlias(["8f", "8bf"], commands)