+++ changes.md
@@ -1,3 +1,14 @@
+# 1.5.3
+
+## Additions
+* nicer [profile
+* [shop as alias to [allitems
+
+## Changes
+* Improve var=value in commands (no longer needs quotes if there is no space)
+
+---
+
 # 1.5.2
 
 ## Fixes